requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
streamlit>=1.38.0
pypdf>=4.0.0
python-dotenv>=1.0.0
googlesearch-python>=1.2.0
